[DEBT-043-D3] Split cli/portfolio.py → cli/portfolio/ package
- Split cli/portfolio.py (620 LoC) into cli/portfolio/ package with 8 files
- Created: _helpers.py, sync.py, positions.py, pnl_cmd.py, balance.py, history.py, link.py, __init__.py
- Largest file: _helpers.py at 129 LoC (all under 400 LoC ceiling)
- Renamed _format_signed_currency → format_signed_currency (public API)
- Updated test to use new public API name
- Added noqa: TC003 for Path imports (required at runtime for Typer introspection)
- Quality gates: pre-commit passed, pytest (1052 passed)
- Acceptance criteria: DEBT-043 Phase D3 [x]
